requirements.txt
fastapi
uvicorn
flask
openai
python-dotenv

